refactor: Update Webpush domain folder structure
