+++ src/main/java/vom/spring/domain/webpush/controller/WebpushController.java
@@ -1,18 +1,14 @@
-package vom.spring.domain.webpush;
+package vom.spring.domain.webpush.controller;
 
-import lombok.Getter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import vom.spring.domain.webpush.dto.WebpushDto;
+import vom.spring.domain.webpush.service.WebpushService;
 
 import java.util.List;
